Moved comments to its own component file
Added Comment component file
Added JSX reference in Meme component file
Passed user and meme props to comment file
